requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==3.2.4
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.1
pluggy==1.0.0
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
